changed method from get to delete, added exception for Delete Hotel

+++ db/db_hotel.py
@@ -1,6 +1,7 @@
 from sqlalchemy.orm import Session
 from db.models import Dbhotel
 from schemas import HotelBase
+from fastapi import HTTPException, status
 
 def create_hotel(db: Session, request: HotelBase):
     new_hotel = Dbhotel(
@@ -19,6 +20,9 @@ def create_hotel(db: Session, request: HotelBase):
 #delete hotel
 def delete_hotel(db: Session, id: int):
     hotel = db.query(Dbhotel).filter(Dbhotel.id == id).first()
+    if not hotel:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
+            detail=f"Hotel with {id} not found")
     db.delete(hotel)
     db.commit()
     return "ok"
